+++ src/registration/index.ts
@@ -6,8 +6,9 @@ import { Signer } from '../keys/types/internal'
 import { Address } from '../types/internal'
 
 export interface RegistrationParams {
+  /** initial programs associated with the user */
   programData: ProgramInstance[]
-  /** just testing this functionality, not intending to use this as the set program */
+  /** The account authorized to modify programs on behalf of the user. */
   programDeployer?: SS58Address
 }
 
@@ -91,14 +92,7 @@ export default class RegistrationManager extends ExtrinsicBaseClass {
     const registerTx = this.substrate.tx.registry.register(
       programDeployer,
       keyVisibility,
-      programData.map((programInfo) => {
-        return {
-          program_pointer: programInfo.program_pointer,
-          program_config: Array.from(
-            Buffer.from(JSON.stringify(programInfo.program_config))
-          ),
-        }
-      })
+      programData.map(this.#formatProgramInfo)
     )
     // @ts-ignore: next line
     // Send the registration transaction and wait for the result.
@@ -158,5 +152,13 @@ export default class RegistrationManager extends ExtrinsicBaseClass {
         })
       })
     })
+  }
+
+  #formatProgramInfo (programInfo): ProgramInstance {
+    const program: ProgramInstance = { program_pointer: programInfo.program_pointer }
+    if (programInfo.program_config) program.program_config = Array.from(
+      Buffer.from(JSON.stringify(programInfo.program_config))
+    )
+    return program
   }
 }