+++ tsup.config.ts
@@ -3,12 +3,14 @@ import { defineConfig } from 'tsup'
 export default defineConfig((options) => {
   return {
     entry: [
-      `src/index.ts`,
+      'src/index.ts',
       'src/keys/index.ts',
       'src/keys/utils.ts',
       'src/utils/index.ts',
       'src/utils/crypto/index.ts',
     ],
+    treeshake: true,
+    skipNodeModulesBundle: true,
     replaceNodeEnv: true,
     format: ['esm', 'cjs'],
     dts: true,
@@ -17,6 +19,5 @@ export default defineConfig((options) => {
     clean: true,
     target: 'es2022',
     minify: options.minify,
-    watch: options.watch,
   }
 })